Lift `rateLoginInput` heuristics from `ratePasswordInput`
My use case is: for some reason, the email field is not detected on
Discord. This means I have to either 1) type in my email or 2) inspect
element and change the empty `name=""` to something from my login input
list. This change directly copies the implementation of
`ratePasswordInput`. I tested on Discord, GitHub, and Google login pages
and everything still appears to work.