HTTP Auth: limit the number of popups
This change prevents from opening a large number of popups for embedded
objects requiring HTTP Auth (e.g., public or cached page data).

We do so by searching the currentAuths array for existing entries that
share the same Root URI and the same realm. If one is found, we don't
open a new popup, but return the Promise associated with the first auth
instead.

fixes: #343, #322

Signed-off-by: Olivier Mehani <shtrom@ssji.net>